Breaking out list.expandColumns and list.selectColumns
