+++ lib/list/isReserved.js
@@ -23,6 +23,6 @@ var reservedPaths = [
 
 function isReserved (path) {
 	return reservedPaths.indexOf(path) >= 0;
-};
+}
 
 module.exports = isReserved;
